Made the interface to 'Task' generic
A 'Task' is now a generic interface. A new class called ProcessTask implements
the functionality of the subprocess execution that previously occurred in
Task. Another class called FunctionTask was added that runs Python functions
as tasks. The interface is generic, so users can define their own if they want
to.