Abstracted the resource interface.
The Resource class is now abstract. The ResourceManager now delegates
all responsibilities to its Resource objects. Specific resource classes
can be created to perform custom resource management.

The old functionality of the Resource class is now in the CounterResource
class, which just counts the amount of current resource against the total
amount.