Implemented ProcessTask prefuncs.
ProcessTask is now fitted with the ability to have list of preexec_fn's that can
is executed in the subprocess before the actual command is executed. This is being
used to set the memory limit using resource.setrlimit instead of the old way of
modifying the command with ulimit -v.